Added a Python scraping prototype
Can be scaled to scrape a single website in a loop or cycle. 

Could also be set to do so with the push of a button, to have control over timing any single scrape which will be. 
Another possibility: After metrics are scraped for one pitch and the pitch is tagged with metrics, the scraper won't print until the next batch of scraped content are numbers and not dashes or blank marks.